# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/reports/request": {
      "post": {
        "operationId": "ReportsController_generate",
        "summary": "Generar un informe de recogidas (PDF Blue Resilience) y subirlo a S3",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequestDto"
              },
              "examples": {
                "monthlyAll": {
                  "summary": "Informe mensual (todas las campañas)",
                  "value": {
                    "type": "monthly",
                    "period": {
                      "preset": "month"
                    },
                    "scope": {
                      "campaign": "all"
                    },
                    "detail": "standard",
                    "language": "en",
                    "format": "pdf"
                  }
                },
                "annualAll": {
                  "summary": "Informe anual",
                  "value": {
                    "type": "annual",
                    "period": {
                      "preset": "year"
                    },
                    "scope": {
                      "campaign": "all"
                    }
                  }
                },
                "campaignC2": {
                  "summary": "Informe de campaña (c2 → Badalona)",
                  "value": {
                    "type": "campaign",
                    "period": {
                      "preset": "all"
                    },
                    "scope": {
                      "campaign": "c2"
                    }
                  }
                },
                "customRange": {
                  "summary": "Periodo personalizado",
                  "value": {
                    "type": "custom",
                    "period": {
                      "start": "2025-01-01",
                      "end": "2025-12-31"
                    },
                    "scope": {
                      "campaign": "all"
                    }
                  }
                },
                "evidenceReport": {
                  "summary": "Informe de evidencias",
                  "value": {
                    "type": "evidence",
                    "period": {
                      "preset": "all"
                    },
                    "scope": {
                      "campaign": "all"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "Analytics API of ONDAs DataSpace",
    "description": "Documentación del API de analíticas. Los nombres de los campos se mantienen en inglés. POST /v1/auth/login devuelve JWT; usar cabecera Authorization: Bearer (token) en POST /v1/analyses/run.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "portal-jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token de POST /v1/auth/login"
      }
    },
    "schemas": {
...
      "ReportPeriodDto": {
        "type": "object",
        "properties": {
          "preset": {
            "type": "string",
            "enum": [
              "month",
              "year",
              "2024",
              "all"
            ]
          },
          "start": {
            "type": "string",
            "example": "2025-03-01"
          },
          "end": {
            "type": "string",
            "example": "2025-05-31"
          }
        }
      },
      "ReportScopeDto": {
        "type": "object",
        "properties": {
          "campaign": {
            "type": "string",
            "example": "all",
            "description": "'all' | campaign id (c1..c4)"
          },
          "entity": {
            "type": "string",
            "example": "auto"
          }
        }
      },
      "ReportIncludeDto": {
        "type": "object",
        "properties": {
          "kpis": {
            "type": "boolean",
            "default": true
          },
          "map": {
            "type": "boolean",
            "default": true
          },
          "charts": {
            "type": "boolean",
            "default": true
          },
          "cleanupsList": {
            "type": "boolean",
            "default": true
          },
          "evidence": {
            "type": "boolean",
            "default": true
          },
          "plasticTypes": {
            "type": "boolean",
            "default": true
          },
          "ondas": {
            "type": "boolean",
            "default": false
          },
          "impactIndex": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "ReportRequestDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "monthly",
              "annual",
              "campaign",
              "location",
              "evidence",
              "custom"
            ],
            "example": "monthly"
          },
          "period": {
            "$ref": "#/components/schemas/ReportPeriodDto"
          },
          "scope": {
            "$ref": "#/components/schemas/ReportScopeDto"
          },
          "detail": {
            "type": "string",
            "enum": [
              "summary",
              "standard",
              "detailed"
            ],
            "default": "standard"
          },
          "language": {
            "type": "string",
            "enum": [
              "en",
              "es",
              "fr"
            ],
            "default": "en"
          },
          "format": {
            "type": "string",
            "enum": [
              "pdf",
              "xlsx"
            ],
            "default": "pdf",
            "description": "xlsx is coerced to pdf"
          },
          "include": {
            "$ref": "#/components/schemas/ReportIncludeDto"
          }
        },
        "required": [
          "type"
        ]
      },
      "ReportResponseDto": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "example": "rep_ab12cd34"
          },
          "status": {
            "type": "string",
            "example": "ready"
          },
          "name": {
            "type": "string",
            "example": "Monthly cleanup report — June 2026"
          },
          "type": {
            "type": "string",
            "example": "monthly"
          },
          "period": {
            "type": "string",
            "example": "June 2026"
          },
          "generatedAt": {
            "type": "string",
            "example": "2026-06-18T10:32:00Z"
          },
          "format": {
            "type": "string",
            "example": "pdf"
          },
          "size": {
            "type": "string",
            "example": "0.4 MB"
          },
          "downloadUrl": {
            "type": "string",
            "example": "https://universalplastic-sedia.s3.eu-central-1.amazonaws.com/public/mediterraneo/universal_plastic/reports/rep_ab12cd34.pdf"
          }
        },
        "required": [
          "requestId",
          "status",
          "name",
          "type",
          "period",
          "generatedAt",
          "format",
          "size",
          "downloadUrl"
        ]
      },
...
    }
  }
}